+++ src/composables/useFormFieldIds.ts
@@ -1,0 +1,19 @@
+import { computed, type ComputedRef, useAttrs, useSlots } from 'vue'
+
+export const useFormFieldIds = (elementId: ComputedRef<string>) => {
+  const attrs = useAttrs()
+  const slots = useSlots()
+
+  const validationMessageId = computed(() => `${elementId.value}-validation`)
+  const helperId = computed(() => `${elementId.value}-helper`)
+
+  const ariaDescribedby = computed(() => {
+    const ids: string[] = []
+    if (attrs['aria-describedby']) ids.push(String(attrs['aria-describedby']))
+    if (slots.validationMessage) ids.push(validationMessageId.value)
+    if (slots.helper) ids.push(helperId.value)
+    return ids.length ? ids.join(' ') : undefined
+  })
+
+  return { ariaDescribedby, helperId, validationMessageId }
+}
